+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-10-24
+
+### Added
+- `ToMap[T, K](List[T], func(T) K)` - Convert list to map using key function
+- `GroupBy[T, K](List[T], func(T) K)` - Group list items by key into Map[K, List[T]]
+
 ## [1.0.1] - 2025-10-24
 
 ### Changed